refactor: extract encoding state from app
